mem: Update ideal dram

+++ mem/dram/ideal.hh
@@ -29,8 +29,8 @@ class Ideal : public AbstractDRAM {
   Ideal(ObjectData &);
   ~Ideal();
 
-  void read(uint64_t, uint64_t, Event, uint64_t) override;
-  void write(uint64_t, uint64_t, Event, uint64_t) override;
+  void read(uint64_t, uint32_t, Event, uint64_t) override;
+  void write(uint64_t, uint32_t, Event, uint64_t) override;
 
   void createCheckpoint(std::ostream &) const noexcept override;
   void restoreCheckpoint(std::istream &) noexcept override;
